refactor(whatsrust): extract message models
